Split import into import and import_helm
.. to make way for import_image.
